Move hanging ydb test to the separate file
